[backend] truncate global score success percentage to have only one decimal

+++ openbas-api/src/main/java/io/openbas/rest/scenario/response/GlobalScoreBySimulationEndDate.java
@@ -7,4 +7,4 @@ import java.time.Instant;
 public record GlobalScoreBySimulationEndDate(
     @JsonProperty("simulation_end_date") @NotNull Instant simulationEndDate,
     @JsonProperty("global_score_success_percentage") @NotNull
-        double globalScoreSuccessPercentage) {}
+        float globalScoreSuccessPercentage) {}
